** Major Changes (11 May 2015): Migrated the majority of code off of the player character and into seperate components. Removed Old maps Add a WIP blocked out Tower map. (Feel free to change assets or remodel)
Upcoming:
The Basic UI is hooked back up now as far as turning windows on and off. Still need to fix a bug where the character screen opens instead of the inventory.  I still need to hook inputs back up in the UI system to account for new changes. I will work on that tomorrow.
More to come with the new component system.
Will be  cleaning out legacy blueprints in the next couple of days.